modules/requirements.txt
Flask
pandas
matplotlib
numpy
Pillow
pygame
scipy
seaborn